Clean up notes and old functions
