+++ client/src/components/related/ComparisonModal.jsx
@@ -1,0 +1,65 @@
+import React, {useState, useEffect} from 'react';
+import RelatedLib from './assets/Stylesheet.jsx';
+import Modal from 'react-modal';
+import Table from './assets/table.jsx';
+import StyleLib from '../../assets/Stylesheet.jsx';
+
+Modal.setAppElement('#root');
+
+const ComparisonModal = ({products, isOpen, onClose}) => {
+  const [productList, setProductList] = useState([]);
+  const [features, setFeatures] = useState([]);
+
+  const getAllFeatures = (product1, product2) => {
+    let features = new Set();
+    let combined = [
+      ...products[0].details.features,
+      ...products[1].details.features
+    ];
+    for (let feat of combined) {
+      features.add(feat.feature)
+    }
+    setFeatures([...features]);
+  }
+
+  const formatData = (product) => {
+    let newDataFormat = {
+      'name': product.details.name,
+      'id': product.details.id,
+      'features': getCharacteristics(product)
+    };
+    return newDataFormat;
+  }
+
+  const getCharacteristics = (product) => {
+    let characteristics = {};
+    for (let feat of product.details.features) {
+      characteristics[feat.feature] = feat.value;
+    }
+    return characteristics;
+  }
+
+  const formatAll = (products) => {
+    let data = [];
+    getAllFeatures(products[0], products[1])
+    for (let product of products) {
+      data.push(formatData(product));
+    }
+    setProductList(data);
+  }
+
+  return (
+    <Modal
+    isOpen={isOpen}
+    onAfterOpen={() => formatAll(products)}
+    onRequestClose={onClose}
+    >
+      <span>Comparing</span>
+      <div className='modal-body'>
+        <Table currentProduct={productList[0]} selectedProduct={productList[1]} features={features}/>
+      </div>
+    </Modal>
+  )
+}
+
+export default ComparisonModal;
